Add def file to clean CreateMatrixfromModelEFConSPHERE.py
